Started writing code to make STM32 send datetime and temp data to ESP32 when it requests a photo to be taken

+++ STM32/Library/Src/ds18b20.c
@@ -45,7 +45,7 @@
 typedef struct ds18b20_t {
     uint8_t id;
     uint64_t rom;
-    uint16_t decimal;
+    uint8_t decimal;
     uint16_t fraction;
     uint8_t sign; // 0 = positive number, 1 = negative number
 } ds18b20_t;
